pyproject.toml
[tool.poetry]
name = "telegram_bot_base"
version = "0.1.0"
description = ""
authors = ["Georgiy <george27khan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.21.0"
aiogram = "2.23.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
